+++ app/src/main/java/com/example/timetracker/data/ActiveState.kt
@@ -1,5 +1,0 @@
-package com.example.timetracker.data
-
-enum class ActiveState {
-    IDLE, WORK, SELF, SLEEP
-}